trivial: promote state_key.rs to folder
