+++ aptos-move/aptos-gas-profiling/src/aggregate.rs
@@ -6,7 +6,6 @@ use crate::{
     render::{Render, TableKey},
 };
 use aptos_gas_algebra::{GasQuantity, GasScalingFactor, InternalGas};
-use aptos_types::state_store::state_key::StateKeyInner;
 use std::collections::{btree_map, BTreeMap};
 
 /// Represents an aggregation of execution gas events, including the count and total gas costs for each type of event.
@@ -113,7 +112,7 @@ impl ExecutionAndIOCosts {
         }
 
         for write in &self.write_set_transient {
-            use StateKeyInner::*;
+            use aptos_types::state_store::state_key::inner::StateKeyInner::*;
 
             let key = match write.key.inner() {
                 AccessPath(ap) => format!("{}", Render(&ap.get_path())),
