+++ aptos-move/aptos-vm-types/src/change_set.rs
@@ -19,7 +19,7 @@ use aptos_types::{
     contract_event::ContractEvent,
     delayed_fields::PanicError,
     state_store::{
-        state_key::{StateKey, StateKeyInner},
+        state_key::{inner::StateKeyInner, StateKey},
         state_value::StateValueMetadata,
     },
     transaction::ChangeSet as StorageChangeSet,
