+++ storage/aptosdb/src/db/mod.rs
@@ -57,8 +57,7 @@ use aptos_types::{
     },
     state_proof::StateProof,
     state_store::{
-        state_key::StateKey,
-        state_key_prefix::StateKeyPrefix,
+        state_key::{prefix::StateKeyPrefix, StateKey},
         state_storage_usage::StateStorageUsage,
         state_value::{StateValue, StateValueChunkWithProof},
         table::{TableHandle, TableInfo},
