+++ storage/aptosdb/src/schema/state_value/mod.rs
@@ -21,7 +21,10 @@ use aptos_schemadb::{
     schema::{KeyCodec, SeekKeyCodec, ValueCodec},
 };
 use aptos_types::{
-    state_store::{state_key::StateKey, state_key_prefix::StateKeyPrefix, state_value::StateValue},
+    state_store::{
+        state_key::{prefix::StateKeyPrefix, StateKey},
+        state_value::StateValue,
+    },
     transaction::Version,
 };
 use byteorder::{BigEndian, ReadBytesExt, WriteBytesExt};
