+++ types/src/state_store/mod.rs
@@ -19,7 +19,6 @@ use std::{collections::HashMap, ops::Deref};
 pub mod errors;
 pub mod in_memory_state_view;
 pub mod state_key;
-pub mod state_key_prefix;
 pub mod state_storage_usage;
 pub mod state_value;
 pub mod table;
