+++ types/src/state_store/state_key/inner.rs
@@ -1,0 +1,115 @@
+// Copyright © Aptos Foundation
+// SPDX-License-Identifier: Apache-2.0
+
+use crate::{access_path::AccessPath, state_store::table::TableHandle};
+use aptos_crypto::{
+    hash::{CryptoHash, CryptoHasher},
+    HashValue,
+};
+use aptos_crypto_derive::CryptoHasher;
+use num_derive::{FromPrimitive, ToPrimitive};
+use serde::{Deserialize, Serialize};
+use std::{
+    fmt,
+    fmt::{Debug, Formatter},
+};
+use thiserror::Error;
+
+#[repr(u8)]
+#[derive(Clone, Debug, FromPrimitive, ToPrimitive)]
+pub enum StateKeyTag {
+    AccessPath,
+    TableItem,
+    Raw = 255,
+}
+
+impl CryptoHash for StateKeyInner {
+    type Hasher = StateKeyInnerHasher;
+
+    fn hash(&self) -> HashValue {
+        let mut state = Self::Hasher::default();
+        state.update(
+            self.encode()
+                .expect("Failed to serialize the state key")
+                .as_ref(),
+        );
+        state.finish()
+    }
+}
+
+/// Error thrown when a [`StateKey`] fails to be deserialized out of a byte sequence stored in physical
+/// storage, via [`StateKey::decode`].
+#[derive(Debug, Error)]
+pub enum StateKeyDecodeErr {
+    /// Input is empty.
+    #[error("Missing tag due to empty input")]
+    EmptyInput,
+
+    /// The first byte of the input is not a known tag representing one of the variants.
+    #[error("lead tag byte is unknown: {}", unknown_tag)]
+    UnknownTag { unknown_tag: u8 },
+
+    #[error("Not enough bytes: tag: {}, num bytes: {}", tag, num_bytes)]
+    NotEnoughBytes { tag: u8, num_bytes: usize },
+
+    #[error(transparent)]
+    BcsError(#[from] bcs::Error),
+}
+
+#[derive(Clone, CryptoHasher, Eq, PartialEq, Serialize, Deserialize, Ord, PartialOrd, Hash)]
+#[cfg_attr(any(test, feature = "fuzzing"), derive(proptest_derive::Arbitrary))]
+#[serde(rename = "StateKey")]
+pub enum StateKeyInner {
+    AccessPath(AccessPath),
+    TableItem {
+        handle: TableHandle,
+        #[serde(with = "serde_bytes")]
+        key: Vec<u8>,
+    },
+    // Only used for testing
+    #[serde(with = "serde_bytes")]
+    Raw(Vec<u8>),
+}
+
+impl StateKeyInner {
+    /// Serializes to bytes for physical storage.
+    pub fn encode(&self) -> anyhow::Result<Vec<u8>> {
+        let mut out = vec![];
+
+        let (prefix, raw_key) = match self {
+            StateKeyInner::AccessPath(access_path) => {
+                (StateKeyTag::AccessPath, bcs::to_bytes(access_path)?)
+            },
+            StateKeyInner::TableItem { handle, key } => {
+                let mut bytes = bcs::to_bytes(&handle)?;
+                bytes.extend(key);
+                (StateKeyTag::TableItem, bytes)
+            },
+            StateKeyInner::Raw(raw_bytes) => (StateKeyTag::Raw, raw_bytes.to_vec()),
+        };
+        out.push(prefix as u8);
+        out.extend(raw_key);
+        Ok(out)
+    }
+}
+
+impl Debug for StateKeyInner {
+    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
+        match self {
+            StateKeyInner::AccessPath(ap) => {
+                write!(f, "StateKey::{:?}", ap)
+            },
+            StateKeyInner::TableItem { handle, key } => {
+                write!(
+                    f,
+                    "StateKey::TableItem {{ handle: {:x}, key: {} }}",
+                    handle.0,
+                    hex::encode(key),
+                )
+            },
+            StateKeyInner::Raw(bytes) => {
+                write!(f, "StateKey::Raw({})", hex::encode(bytes),)
+            },
+        }
+    }
+}
